Reworked Items
All the items of the player are now in Items.
The Location attribute of Item is not just BodyLocation anymore,
it contains the position in the inventory along with it's width and height.
The Container indicates where the Item is stored (Equipment, Belt, Cube, Inventory, Stash)

Changed GetItemDescription to take a D2Unit (item) instead of Item and made it public
Added Location for hireling items